Event network implementation, structure change, running to completion

+++ maxpool.h
@@ -14,7 +14,7 @@ private:
 public:
 	MaxPool() {}
 	MaxPool(unsigned k, unsigned stri);
-	vector<vector<vector<double>>> feedForward(vector<vector<vector<double>>> &input);
+	vector<vector<vector<float>>> feedForward(vector<vector<vector<float>>> &input);
 };
 MaxPool::MaxPool(unsigned k, unsigned stri) {
 	stride = stri;
@@ -23,15 +23,15 @@ MaxPool::MaxPool(unsigned k, unsigned stri) {
 
 
 
-vector<vector<vector<double>>> MaxPool::feedForward(vector<vector<vector<double>>> &input) {
-	vector<vector<vector<double>>> net_out;
+vector<vector<vector<float>>> MaxPool::feedForward(vector<vector<vector<float>>> &input) {
+	vector<vector<vector<float>>> net_out;
 	for(int i = 0; i < input.size(); i++) {
-		vector<vector<double>> temp_net_out;
+		vector<vector<float>> temp_net_out;
 		temp_net_out.clear();
         for (int y = 0; y <= input[i].size() - kern; y += stride) {
             temp_net_out.push_back({});
             for (int x = 0; x <= input[i][y].size() - kern; x += stride) {
-                double out = input[i][y][x];
+                float out = input[i][y][x];
                 for (int fY = 0; fY < kern; fY++) {
                     for (int fX = 0; fX < kern; fX++) {
                             if (y + fY < input[i].size() && x + fX < input[i][y].size()) {
